Prove that Do is idempotent*
*If the caller has implemented the memory layer (`CheckMemory` and
`Remember`) correctly.

This should give us confidence that a function using the `Do` trait,
such as the `hbit_herc20` one we use to execute Hbit<->Herc20 swaps,
does not lead us to executing the same action twice, even if we were
to call it more than once, as could be the case if we were respawning
a swap.
