+++ Cargo.toml
@@ -38,7 +38,6 @@ default-features = false
 version = "0.6"
 
 [dev-dependencies]
-primitive-types = { version = "0.7.2" }
 proptest = "0.10"
 testcontainers = "0.9"
