+++ src/swap/bob.rs
@@ -331,172 +331,3 @@ pub mod watch_only_actor {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::swap::{Do, Next};
-    use comit::{htlc_location, identity, transaction};
-    use primitive_types::U256;
-    use std::{
-        collections::HashMap,
-        sync::{Arc, RwLock},
-    };
-
-    #[derive(Default)]
-    struct MockDatabase {
-        deploy_events: Arc<RwLock<HashMap<SwapId, herc20::Deployed>>>,
-    }
-
-    struct MockBitcoinWallet;
-
-    struct MockEthereumWallet {
-        node: Arc<RwLock<EthereumBlockchain>>,
-    }
-
-    #[derive(Default)]
-    struct EthereumBlockchain {
-        deploy_events: Vec<herc20::Deployed>,
-    }
-
-    struct Lock {
-        secret: Secret,
-        secret_hash: SecretHash,
-    }
-
-    impl Lock {
-        fn new() -> Self {
-            let bytes = b"hello world, you are beautiful!!";
-            let secret = Secret::from(*bytes);
-
-            let secret_hash = SecretHash::new(secret);
-
-            Self {
-                secret,
-                secret_hash,
-            }
-        }
-    }
-
-    struct SwapTimes {
-        start_of_swap: NaiveDateTime,
-        alpha_expiry: Timestamp,
-        beta_expiry: Timestamp,
-    }
-
-    impl SwapTimes {
-        fn live_swap() -> Self {
-            let start_of_swap = Timestamp::now();
-            let beta_expiry = start_of_swap.plus(60 * 60);
-            let alpha_expiry = beta_expiry.plus(60 * 60);
-
-            Self {
-                start_of_swap: start_of_swap.into(),
-                alpha_expiry,
-                beta_expiry,
-            }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl herc20::ExecuteDeploy for MockEthereumWallet {
-        async fn execute_deploy(
-            &self,
-            _params: herc20::Params,
-        ) -> anyhow::Result<herc20::Deployed> {
-            let deploy_event = herc20::Deployed {
-                transaction: transaction::Ethereum {
-                    hash: ethereum::Hash::from([0u8; 32]),
-                    to: None,
-                    value: U256::from(0u64),
-                    input: Vec::new(),
-                },
-                location: htlc_location::Ethereum::random(),
-            };
-
-            let mut blockchain = self.node.write().unwrap();
-            blockchain.deploy_events.push(deploy_event.clone());
-
-            Ok(deploy_event)
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl BetaLedgerTime for MockEthereumWallet {
-        async fn beta_ledger_time(&self) -> anyhow::Result<Timestamp> {
-            Ok(Timestamp::now())
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl db::Load<herc20::Deployed> for MockDatabase {
-        async fn load(&self, swap_id: SwapId) -> anyhow::Result<Option<herc20::Deployed>> {
-            let deploy_events = self.deploy_events.read().unwrap();
-
-            Ok(deploy_events.get(&swap_id).cloned())
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl db::Save<herc20::Deployed> for MockDatabase {
-        async fn save(
-            &self,
-            deploy_event: herc20::Deployed,
-            swap_id: SwapId,
-        ) -> anyhow::Result<()> {
-            let mut deploy_events = self.deploy_events.write().unwrap();
-            deploy_events.insert(swap_id, deploy_event);
-
-            Ok(())
-        }
-    }
-
-    #[tokio::test]
-    async fn herc20_deploy_is_idempotent() {
-        let ethereum_blockchain = Arc::new(RwLock::new(EthereumBlockchain::default()));
-        let ethereum_wallet = MockEthereumWallet {
-            node: Arc::clone(&ethereum_blockchain),
-        };
-
-        let db = MockDatabase::default();
-
-        let swap_id = SwapId::random();
-
-        let Lock { secret_hash, .. } = Lock::new();
-
-        let SwapTimes {
-            start_of_swap,
-            beta_expiry,
-            ..
-        } = SwapTimes::live_swap();
-
-        let herc20_params = herc20::params(
-            secret_hash,
-            ethereum::ChainId::regtest(),
-            identity::Ethereum::random(),
-            identity::Ethereum::random(),
-            ethereum::Address::random(),
-            beta_expiry,
-        );
-
-        let bob = WalletBob {
-            alpha_wallet: MockBitcoinWallet,
-            beta_wallet: ethereum_wallet,
-            db,
-            private_protocol_details: (),
-            secret_hash,
-            start_of_swap,
-            swap_id,
-        };
-
-        assert!(ethereum_blockchain.read().unwrap().deploy_events.is_empty());
-        let res = bob.r#do(beta_expiry, herc20_params.clone()).await;
-
-        assert!(matches!(res, Ok(Next::Continue(herc20::Deployed { .. }))));
-        assert_eq!(ethereum_blockchain.read().unwrap().deploy_events.len(), 1);
-
-        let res = bob.r#do(beta_expiry, herc20_params).await;
-        assert!(matches!(res, Ok(Next::Continue(herc20::Deployed { .. }))));
-        assert_eq!(ethereum_blockchain.read().unwrap().deploy_events.len(), 1);
-    }
-}
